feat(front/passes/types): organize type pass
